Added property for executing Elasticsearch migrations

+++ src/main/java/com/github/veljko121/gigster/config/ElasticsearchMigrations.java
@@ -1,6 +1,7 @@
 package com.github.veljko121.gigster.config;
 
 import org.slf4j.Logger;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.context.event.ApplicationReadyEvent;
 import org.springframework.context.ApplicationListener;
 import org.springframework.lang.NonNull;
@@ -16,23 +17,32 @@ import lombok.RequiredArgsConstructor;
 public class ElasticsearchMigrations implements ApplicationListener<ApplicationReadyEvent> {
 
     private final Logger logger;
+
     private final GigListingRepository gigListingRepository;
+
     private final IGigsterSearchEngineService gigsterSearchEngineService;
+
+    @Value("${gigster-search-engine.init.mode}") private String INIT_MODE; 
 
     @Override
     public void onApplicationEvent(@NonNull ApplicationReadyEvent event) {
-        logger.info("Starting Elasticsearch migrations.");
-        try {
-            gigsterSearchEngineService.deleteAllGigListings();
-            logger.info("Deleted all gig-listings from Elasticsearch.");
-            var gigListings = gigListingRepository.findAll();
-            for (var gigListing : gigListings) {
-                gigsterSearchEngineService.createGigListing(gigListing);
+        if (INIT_MODE.equals("always")) {
+            logger.info("Starting Elasticsearch migrations.");
+            try {
+                gigsterSearchEngineService.deleteAllGigListings();
+                logger.info("Deleted all gig-listings from Elasticsearch.");
+                var gigListings = gigListingRepository.findAll();
+                for (var gigListing : gigListings) {
+                    gigsterSearchEngineService.createGigListing(gigListing);
+                }
+                logger.info("Migrated all gig-listings to Elasticsearch.");
+                logger.info("Elasticsearch migrations successful.");
+            } catch (Exception e) {
+                logger.error("Error when migrating data to Elasticsearch.");
             }
-            logger.info("Migrated all gig-listings to Elasticsearch.");
-            logger.info("Elasticsearch migrations successful.");
-        } catch (Exception e) {
-            logger.error("Error when migrating data to Elasticsearch.");
+        }
+        else {
+            logger.info("Skipping Elasticsearch migrations.");
         }
     }
